Creates a single NCCS-specific Spack branch
NCCS-specific changes should not be made to the upstream tracked Spack
packages. Instead, NCCS-specific package alterations should be placed in
a separate repo that will not conflict with merges from upstream.

The only allowable merge conflicts should be in Spack python modules
that must be patched to work in our environment. These changes should be
generalized and pushed upstream in order to improve Spack for everyone.

Any changes to this branch that conflict with upstream develop will be
rebased out once they have been merged upstream.

Configuration files for each NCCS host are kept under
`$spack/etc/spack/hosts/{hostname}` and the appropriate config is linked
to by the `sw-olcf` builder script when a copy of spack is deployed. In
this way, changes made to a system's spack configuration can be made
in-place. To be deployed, the changes to a host's Spack config and any
custom packages must be committed and merged into this branch first.